+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/naming/abbreviationaswordinname/InputAbbreviationAsWordInNameType2.java
@@ -106,28 +106,3 @@ class FIleNameFormatException2 extends Exception {
         super(e);
     }
 }
-
-class StateX2 {
-    int userID;
-    int scaleX, scaleY, scaleZ;
-
-    int getScaleX() {
-        return this.scaleX;
-    }
-}
-
-@interface Annotation12 {
-    String VALUE = "value"; // in @interface this is final/static
-}
-
-@interface Annotation22 {
-    static String VALUE = "value"; // in @interface this is final/static
-}
-
-@interface Annotation32 {
-    final String VALUE = "value"; // in @interface this is final/static
-}
-
-@interface Annotation42 {
-    final static String VALUE = "value"; // in @interface this is final/static
-}
