+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/naming/abbreviationaswordinname/InputAbbreviationAsWordInNameType3.java
@@ -106,28 +106,3 @@ class FIleNameFormatException3 extends Exception {
         super(e);
     }
 }
-
-class StateX3 {
-    int userID;
-    int scaleX, scaleY, scaleZ;
-
-    int getScaleX() {
-        return this.scaleX;
-    }
-}
-
-@interface Annotation13 {
-    String VALUE = "value"; // in @interface this is final/static
-}
-
-@interface Annotation23 {
-    static String VALUE = "value"; // in @interface this is final/static
-}
-
-@interface Annotation33 {
-    final String VALUE = "value"; // in @interface this is final/static
-}
-
-@interface Annotation43 {
-    final static String VALUE = "value"; // in @interface this is final/static
-}
